# Conflicts:
#	service/main_service/src/main/java/com/sustech/main_service/controller/ArticleController.java
#	service/main_service/src/main/java/com/sustech/main_service/controller/TopicController.java
#	service/main_service/src/main/java/com/sustech/main_service/entity/Topic.java
#	service/main_service/src/main/java/com/sustech/main_service/mapper/TopicMapper.java
#	service/main_service/src/main/java/com/sustech/main_service/service/TopicService.java
#	service/main_service/src/main/java/com/sustech/main_service/service/impl/TopicServiceImpl.java

+++ service/main_service/src/main/java/com/sustech/main_service/entity/Order.java
@@ -2,18 +2,18 @@ package com.sustech.main_service.entity;
 import lombok.Data;
 @Data
 public class Order {
-    int id;
+    Long id;
     String name = "";
     String picture = "";
     String seller = "";
     String buyer = "";
     String seller_contact = "";
     String buyer_contact = "";
-    int price = 0;
+    Integer price = 0;
     //0是购买者发布
-    int status = 0;
+    Integer status = 0;
     //0是无效
-    int valid = 0;
+    Integer valid = 0;
     String gmt_create = "";
     String gmt_modified = "";
 }
